# internal/api: submit service default rates in /v0.4 endpoint