This change fixes a problem where integrations have trouble matching
against the map of rates sent by the agent. For example, when a trace
has no "env" tag set, the agent will use the default env (which is
"none" or whatever is in "datadog.yaml"). Tracers have no knowledge of
this default so they will not be able to match against the map of rates.

To further explain, if the agent sends the map:

    service:mongo,env:none => 0.5

A tracer will try to use the key `service:mongo,env:` because it will
have no env information, failing to match against the map.

Using this change, tracers can use correct defaults.