@gbbr 2018-12-05
I’m fine with either. I like yours because it requires no changes to the tracers but the down side is that the duplication is confusing, it adds network overhead (not sure if this is a real concern) and the implementation in the agent would not be very clean. 

I’d like to hear from the other reviewers too.