@LotharSee 2018-12-05
@willgittoes-dd A single Agent can receive traces from different `env` (where the `env` is set as a span tag).

I agree with @rochdev 's idea to simply duplicate the entries: also return `service:A,env:` with the same value as `service:A,env:%DEFAULT_ENV`. It is both legitimate (fair to cover both cases) and nicer as it requires no Agent change.