@gbbr 2018-12-06
I've implemented it as suggested. It needed a small refactor to avoid hacky assumption making string matching. PTAL.