@gbbr 2018-12-06
Current benchmark for getting all rates (number represents number of entries):
```
name               old time/op    new time/op    delta
RateByService/1-4     225ns ± 0%     366ns ± 0%    +62.67%
RateByService/2-4     273ns ± 0%     738ns ± 0%   +170.33%
RateByService/3-4     288ns ± 0%    1159ns ± 0%   +302.43%
RateByService/4-4     338ns ± 0%    1381ns ± 0%   +308.58%
RateByService/5-4     351ns ± 0%    1794ns ± 0%   +411.11%
RateByService/6-4     370ns ± 0%    2131ns ± 0%   +475.95%

name               old alloc/op   new alloc/op   delta
RateByService/1-4      256B ± 0%      272B ± 0%     +6.25%
RateByService/2-4      256B ± 0%      368B ± 0%    +43.75%
RateByService/3-4      256B ± 0%      480B ± 0%    +87.50%
RateByService/4-4      256B ± 0%      544B ± 0%   +112.50%
RateByService/5-4      256B ± 0%      656B ± 0%   +156.25%
RateByService/6-4      256B ± 0%      929B ± 0%   +262.89%

name               old allocs/op  new allocs/op  delta
RateByService/1-4      2.00 ± 0%      3.00 ± 0%    +50.00%
RateByService/2-4      2.00 ± 0%      8.00 ± 0%   +300.00%
RateByService/3-4      2.00 ± 0%     13.00 ± 0%   +550.00%
RateByService/4-4      2.00 ± 0%     16.00 ± 0%   +700.00%
RateByService/5-4      2.00 ± 0%     21.00 ± 0%   +950.00%
RateByService/6-4      2.00 ± 0%     24.00 ± 0%  +1100.00%
``
Trying to improve in next commits.